Rename FIFA folder on github
