+++ components/infobox/wikis/easportsfc/infobox_company_custom.lua
@@ -1,32 +1,31 @@
 ---
 -- @Liquipedia
--- wiki=fifa
--- page=Module:Infobox/Person/Player/Custom
+-- wiki=easportsfc
+-- page=Module:Infobox/Company/Custom
 --
 -- Please see https://github.com/Liquipedia/Lua-Modules to contribute
 --
 
 local Class = require('Module:Class')
 local Lua = require('Module:Lua')
-local Page = require('Module:Page')
 
+local Company = Lua.import('Module:Infobox/Company')
 local Injector = Lua.import('Module:Infobox/Widget/Injector')
-local Player = Lua.import('Module:Infobox/Person')
 
 local Widgets = require('Module:Infobox/Widget/All')
 local Cell = Widgets.Cell
 
----@class FifaInfoboxPlayer: Person
-local CustomPlayer = Class.new(Player)
+---@class FifaCompanyInfobox: CompanyInfobox
+local CustomCompany = Class.new(Company)
 local CustomInjector = Class.new(Injector)
 
 ---@param frame Frame
 ---@return Html
-function CustomPlayer.run(frame)
-	local player = CustomPlayer(frame)
-	player:setWidgetInjector(CustomInjector(player))
+function CustomCompany.run(frame)
+	local company = CustomCompany(frame)
+	company:setWidgetInjector(CustomInjector(company))
 
-	return player:createInfobox()
+	return company:createInfobox()
 end
 
 ---@param id string
@@ -34,12 +33,11 @@ end
 ---@return Widget[]
 function CustomInjector:parse(id, widgets)
 	local args = self.caller.args
-
-	if id == 'custom' then
-		table.insert(widgets, Cell{name = 'Agency', content = {Page.makeInternalLink(args.agency)}})
+	if id == 'parent' then
+		table.insert(widgets, Cell{name = 'Focus', content = {args.focus}})
 	end
 
 	return widgets
 end
 
-return CustomPlayer
+return CustomCompany
